Moved App.tsx out of src so expo start will work
