+++ __tests__/App-test.tsx
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import App from '../src/App';
+import App from '../App';
 import renderer from 'react-test-renderer';
 
 it('renders the home screen', async () => {
